# Performance improvements - allocs and times