--- a/src/HCubature.jl
+++ b/src/HCubature.jl
@@ -45,6 +45,7 @@ function (::Trivial)(f, a::SVector{0}, b::SVector{0}, norm)
 end
 cubrule(::Val{0}, ::Type{T}) where {T} = Trivial()
 countevals(::Trivial) = 1
+
 struct FirstEval{TI, TE, TΔ, TB, TR}
   I::TI
   E::TE