@stevengj 2022-10-19
I'm wary of adding a hack like this if there's not a clear understanding of why it works…

Is the problem here simply a [lack of specialization on function arguments](https://docs.julialang.org/en/v1/manual/performance-tips/#Be-aware-of-when-Julia-avoids-specializing), that could be fixed by changing `f` to `f::F` and adding `where {F}`?